bring back the player name and color to the settings menu
closes #6380
